Add specs
Would prefer to use Cedar, but since this project was setup as
a pure swift environment, actually getting this to build against
Cedar is a bit of a nightmare. Nothing is exposed for objc and there
is a rampant use of structs. This is totally fine, since it was never
intended to actually be used from an objective-c context, but it's
still frustrating since I was hopeful to use this to prove out some
Swift+Cedar usecases.

Lesson learned, and it's still nice to get more experience with Quick.